added upload-file.js to src/shared in anticipation of further use
