+++ src/server/components/notebook-actions-menu.jsx
@@ -4,29 +4,7 @@ import Menu from './menu'
 import MenuItem from './menu-item'
 import MenuDivider from './menu-divider'
 import DeleteModal from './delete-modal'
-import fetchWithCSRFToken from '../../shared/fetch-with-csrf-token'
-
-function uploadFile(notebookID, successCallback = () => {}) {
-  const filePicker = document.createElement('input')
-  filePicker.type = 'file'
-  filePicker.id = 'file-picker'
-  filePicker.name = 'files[]'
-  filePicker.click();
-  filePicker.addEventListener('change', (evt) => {
-    const file = evt.target.files[0];
-    const formData = new FormData();
-    formData.append('metadata', JSON.stringify({
-      filename: file.name,
-      notebook_id: notebookID,
-    }))
-    formData.append('file', file)
-    fetchWithCSRFToken('/api/v1/files/', {
-      body: formData,
-      method: 'POST',
-    }).then(output => output.json())
-      .then(output => successCallback(output))
-  })
-}
+import uploadFile from '../../shared/upload-file'
 
 export default class NotebookActionsMenu extends React.Component {
   constructor(props) {
